# backend/crates/api/src/auth/authorization.rs
// SPDX-License-Identifier: AGPL-3.0
// Copyright (C) 2024-2026 Breakdown RS Contributors
// Co-authored-by: gpt-5.6-luna (opencode-go)
// Co-authored-by: hy3 (opencode-go)
// Co-authored-by: deepseek-v4-flash (opencode-go)
// Co-authored-by: hy4-preview (opencode-go)
// Co-authored-by: omen-alpha (opencode-go)
// Co-authored-by: glm-5.3-flash (neuralwatt)

//! Authorization policy for the API layer (Section 5, Decision D2/D5).
//!
//! The [`AuthorizationPolicy`] port and [`PolicyDecision`]/[`Action`] types
//! live in `core`; this module provides the concrete, membership-backed
//! implementation and the Axum middleware that enforces it. Authorization is
//! **action-scoped** (the caller's active block, Decision D2) and runs *after*
//! the [`crate::auth::auth_middleware`].

use std::sync::Arc;

use axum::extract::FromRequestParts;
use axum::http::Method;
use axum::middleware::Next;
use axum::response::{IntoResponse, Response};
use breakdown_core::error::DomainError;
use breakdown_core::membership::MembershipRepository;
use breakdown_core::membership::policy::{
    Action, AuthContext, AuthorizationPolicy, PolicyDecision, SeasonAuthContext,
};
use breakdown_core::shared::UserId;

use crate::auth::{ActiveBlock, AuthError, CurrentUser};

/// Membership-backed authorization policy (v1: membership-only).
///
/// Generic over the read-model repository so it can be constructed with the
/// production `MembershipRepositoryImpl` (and remain unit-testable). Exposes
/// the dyn-safe [`AuthorizationPolicy`] port.
pub struct MembershipAuthorizationPolicy<Repo: MembershipRepository> {
    repo: Arc<Repo>,
    /// Deployment-scoped ops bootstrap allowlist (`OPS_ADMIN_SUBS`, issue
    /// #409): trusted OIDC subs that hold ops access even before any
    /// `ops_admin` membership row exists (cold start). Consulted by
    /// [`Self::authorize_ops`]; the membership predicate stays the durable,
    /// delegable grant path.
    ops_admins: Vec<String>,
}

impl<Repo: MembershipRepository> MembershipAuthorizationPolicy<Repo> {
    /// Build a policy that consults `repo` for active-membership (no ops
    /// bootstrap allowlist — use [`Self::with_ops_admins`] to add one).
    pub fn new(repo: Arc<Repo>) -> Self {
        Self {
            repo,
            ops_admins: Vec::new(),
        }
    }

    /// Attach the `OPS_ADMIN_SUBS` bootstrap allowlist (issue #409).
    pub fn with_ops_admins(mut self, subs: Vec<String>) -> Self {
        self.ops_admins = subs;
        self
    }
}

#[async_trait::async_trait]
impl<Repo: MembershipRepository + 'static> AuthorizationPolicy
    for MembershipAuthorizationPolicy<Repo>
{
    async fn authorize(&self, ctx: &AuthContext) -> PolicyDecision {
        match self
            .repo
            .is_active_member(ctx.block_id, ctx.actor.clone())
            .await
        {
            Ok(true) => PolicyDecision::Allow,
            // Any error (or non-member) fails closed to Deny.
            _ => PolicyDecision::Deny,
        }
    }

    async fn authorize_season_result(
        &self,
        ctx: &SeasonAuthContext,
    ) -> Result<PolicyDecision, DomainError> {
        self.repo
            .has_active_costume_role_in_season(ctx.season_id, ctx.actor.clone())
            .await
            .map(|active| {
                if active {
                    PolicyDecision::Allow
                } else {
                    PolicyDecision::Deny
                }
            })
    }

    async fn authorize_credential_role(
        &self,
        actor: &UserId,
    ) -> Result<PolicyDecision, DomainError> {
        self.repo
            .has_active_credential_role(actor.clone())
            .await
            .map(|active| {
                if active {
                    PolicyDecision::Allow
                } else {
                    PolicyDecision::Deny
                }
            })
    }

    async fn authorize_ops(&self, actor: &UserId) -> Result<PolicyDecision, DomainError> {
        // Bootstrap allowlist first (infallible config): deployment operators
        // hold ops access even with no `ops_admin` membership row yet
        // (issue #409 cold start).
        if self.ops_admins.iter().any(|sub| sub == actor.as_str()) {
            return Ok(PolicyDecision::Allow);
        }
        self.repo
            .has_active_ops_role(actor.clone())
            .await
            .map(|active| {
                if active {
                    PolicyDecision::Allow
                } else {
                    PolicyDecision::Deny
                }
            })
    }
}

/// Season-scoped costume-photo authorization policy (v1: derived).
///
/// Grants access if the user holds any costume-dept role
/// (`costume_designer`, `wardrobe_supervisor`, `costume_assistant`) in any
/// `active` block of the target season.
///
/// ## Between-blocks gap (v1 limitation)
///
/// A costumer between contracts (left Block 3, not yet in Block 5 of the
/// same season) loses photo access. This is accepted as correct from a
/// security standpoint — when a user is not on the production, they do not
/// get confidential photos.
///
/// ## v2 evolution (SeasonCrew)
///
/// When users hit the between-blocks gap, an additive `SeasonCrew` aggregate
/// will be introduced. The upgrade path is non-breaking:
/// `authorized = derived-from-active-block OR season-crew-grant`.
/// The `SeasonPhotoAccessPolicy` trait method signature is unchanged — only
/// the impl changes. See ADR-019, D4 for the full evolution plan.
pub struct SeasonPhotoAccessPolicy<Repo: MembershipRepository> {
    repo: Arc<Repo>,
}

impl<Repo: MembershipRepository> SeasonPhotoAccessPolicy<Repo> {
    /// Build a policy that consults `repo` for season-scoped active
    /// costume-dept membership.
    pub fn new(repo: Arc<Repo>) -> Self {
        Self { repo }
    }
}

#[async_trait::async_trait]
impl<Repo: MembershipRepository + 'static> AuthorizationPolicy for SeasonPhotoAccessPolicy<Repo> {
    async fn authorize(&self, _ctx: &AuthContext) -> PolicyDecision {
        // Block-scoped authorization is not applicable to photo access.
        PolicyDecision::Deny
    }

    async fn authorize_season(&self, ctx: &SeasonAuthContext) -> PolicyDecision {
        match self
            .repo
            .has_active_costume_role_in_season(ctx.season_id, ctx.actor.clone())
            .await
        {
            Ok(true) => PolicyDecision::Allow,
            // Any error (or non-costume-role) fails closed to Deny.
            _ => PolicyDecision::Deny,
        }
    }

    async fn authorize_season_result(
        &self,
        ctx: &SeasonAuthContext,
    ) -> Result<PolicyDecision, DomainError> {
        self.repo
            .has_active_costume_role_in_season(ctx.season_id, ctx.actor.clone())
            .await
            .map(|active| {
                if active {
                    PolicyDecision::Allow
                } else {
                    PolicyDecision::Deny
                }
            })
    }
}

/// Shared authorization state for the [`authorize_middleware`].
#[derive(Clone)]
pub struct AuthorizationState {
    policy: Arc<dyn AuthorizationPolicy>,
    /// When `true`, denials return `403`. When `false` (log-only / staged
    /// rollout), every request is allowed but denials are logged.
    enforce: bool,
}

impl AuthorizationState {
    /// Read-only access to the enforcement flag (for diagnostics/logging).
    pub fn enforce(&self) -> bool {
        self.enforce
    }

    /// Production/test state.
    pub fn new(policy: Arc<dyn AuthorizationPolicy>, enforce: bool) -> Self {
        Self { policy, enforce }
    }

    /// Read the enforcement flag from `AUTHZ_ENFORCE` (default: enforce).
    pub fn enforce_from_env() -> bool {
        std::env::var("AUTHZ_ENFORCE")
            .map(|v| v != "false" && v != "0")
            .unwrap_or(true)
    }

    /// Build the authorization state from the environment.
    ///
    /// In dev mode (`DEV_AUTH_SUB` set, i.e. the auth layer is not really
    /// verifying tokens) enforcement defaults to *off* (log-only) so local
    /// development works without seeded membership; set `AUTHZ_ENFORCE=true`
    /// to force enforcement even in dev. Production (no `DEV_AUTH_SUB`)
    /// enforces by default.
    pub fn from_env_or_dev(policy: Arc<dyn AuthorizationPolicy>) -> Self {
        let dev = std::env::var("DEV_AUTH_SUB").is_ok();
        let enforce = if dev {
            std::env::var("AUTHZ_ENFORCE")
                .map(|v| v != "false" && v != "0")
                .unwrap_or(false)
        } else {
            Self::enforce_from_env()
        };
        AuthorizationState::new(policy, enforce)
    }
}

/// How strongly a given route is gated.
#[derive(Debug, PartialEq)]
pub enum Requirement {
    /// Authentication only (no active-block membership required).
    Authenticated,
    /// Requires the caller to be an active member of the active block.
    BlockMember,
}

/// Resolve the authorization requirement for a request path (Section 5.3/5.4).
///
/// Season endpoints and block *creation*/*listing* need only authentication;
/// every other (block-scoped) read or write requires active membership in the
/// active block conveyed by `X-Active-Block` (Decision D2). The active block is
/// not derived from the data being mutated (action-scoped, not data-scoped).
pub fn requirement_for(path: &str) -> Requirement {
    // ADR-021 D1: routes are mounted under `/v1`. Strip the path version so
    // the requirement classification is version-independent (a future `/v2`
    // inherits the same policy until a route is deliberately reclassified).
    let path = path.strip_prefix("/v1").unwrap_or(path);
    // `/series/{id}/membership` is the series-level counterpart of
    // `/seasons/{id}/membership` (issue #535): the membership self-check is
    // tenant-level data the middleware's `X-Active-Block` scope says nothing
    // about, so it needs only an authenticated caller (the handler carries
    // no privileged action). The exception is deliberately NARROW (exactly
    // `{id}/membership`, non-empty id, no deeper segments) so any future
    // `/series/*` route falls back to the default `BlockMember` policy and
    // must be classified explicitly — defense-in-depth against a silently
    // widened allowlist.
    let is_series_membership = path
        .strip_prefix("/series/")
        .and_then(|rest| rest.strip_suffix("/membership"))
        .is_some_and(|id| !id.is_empty() && !id.contains('/'));
    if path.starts_with("/seasons") || is_series_membership || path.starts_with("/settings") {
        return Requirement::Authenticated;
    }
    // `/blocks` covers both `POST` (create + owner bootstrap) and `GET`
    // (list by season) — neither needs an existing active-block membership.
    if path == "/blocks" {
        return Requirement::Authenticated;
    }
    // Photo endpoints — costume photos (`/costumes/{id}/photos…`) and the
    // continuity photos of a scene shoot (`…/scene-shoots/{id}/continuity-photos…`).
    // Both families use season-scoped authorization (SeasonPhotoAccessPolicy /
    // the shooting_day → episode → block → season chain) checked *inside* the
    // handler (see AGENTS.md §7, `// AUTHZ-GATE:`), not by this middleware.
    // `/continuity-photos` is matched explicitly: a bare `contains("/photos")`
    // misses the hyphenated segment.
    if path.contains("/photos") || path.contains("/continuity-photos") {
        return Requirement::Authenticated;
    }

    // AI import endpoints are gated by `Authenticated`; each handler performs
    // handler-internal authorization (costume-dept membership / credential
    // role) and carries an `// AUTHZ-GATE:` comment (mirrors photo handlers).
    if path.starts_with("/ai-import") {
        return Requirement::Authenticated;
    }
    // PDF report endpoints use handler-internal season-scoped auth gates
    // (// AUTHZ-GATE:), just like photo endpoints.
    if path.ends_with(".pdf") && path.contains("/report/") {
        return Requirement::Authenticated;
    }
    // JSON report endpoints (the non-PDF twins of the PDF reports: dispo,
    // shoot-day, soll-ist) carry the same handler-internal season-scoped auth
    // gate, so they are authenticated-only as well (issue #333). They are
    // matched explicitly — not by widening the `.pdf` conjunction above into a
    // `contains("/report/")` disjunction — so the `&&`-vs-`||` mutation test
    // in `crates/api/tests/auth_authorization.rs` stays meaningful and no
    // unrelated `/report/*` route silently drops its block-membership check.
    if path.starts_with("/shooting-days/")
        && path.contains("/report/")
        && (path.ends_with("/dispo") || path.ends_with("/shoot-day") || path.ends_with("/soll-ist"))
    {
        return Requirement::Authenticated;
    }

    // Aggregated report routes (issue #571): gated internally by the
    // handler-internal AUTHZ-GATE over the season membership policy — the
    // same pattern as the day-scoped JSON reports. The season JSON route is
    // covered by the `/seasons` prefix above.
    if path.contains("/report/soll-ist")
        && (path.starts_with("/episodes/") || path.starts_with("/seasons/"))
    {
        return Requirement::Authenticated;
    }

    // Manual "archive now" uses handler-internal AUTHZ-GATE (designer/supervisor).
    if path.ends_with("/report/archive") {
        return Requirement::Authenticated;
    }
    // The series-scoped audit journal (`GET /audit`) is **not** block-scoped:
    // it is filtered by the `project_id` **query parameter**, so the caller's
    // active block (`X-Active-Block`) says nothing about the series whose
    // journal is requested. The handler therefore performs the membership
    // check itself (`MembershipRepository::has_active_membership_in_project`,
    // `// AUTHZ-GATE:`) and returns `403` on denial (issue #342). Classified
    // `Authenticated` like the other handler-gated route families above.
    if path == "/audit" {
        return Requirement::Authenticated;
    }

    // Issue #443 (test-support builds only): the fault-injection control
    // route is authentication-gated, not block-scoped — arming a fault is a
    // test orchestration action, the route only exists in test-support
    // builds, and classification here costs production nothing (no such
    // route is mounted in release binaries).
    if path.starts_with("/__faults") {
        return Requirement::Authenticated;
    }

    // Ops endpoints (issue #409) expose deployment-wide infrastructure state,
    // not block-scoped production data. The handler performs the ops gate
    // itself (`AuthorizationPolicy::authorize_ops` — active `ops_admin`
    // membership in any block or the `OPS_ADMIN_SUBS` bootstrap allowlist,
    // `// AUTHZ-GATE:`) and returns `403` on denial. Classified
    // `Authenticated` like the other handler-gated route families above.
    if path.starts_with("/ops") {
        return Requirement::Authenticated;
    }

    // Everything else (scenes, characters, costumes, episodes, and
    // block detail / time-span updates) is block-scoped.
    // Self-service invitation acceptance: the invitee is *not yet* an active
    // member (they are pending until they accept), so this endpoint cannot
    // require active-block membership. The domain command enforces that the
    // caller actually holds a pending invitation for this block.
    if path.ends_with("/members/accept") {
        return Requirement::Authenticated;
    }

    Requirement::BlockMember
}

/// Axum middleware implementing the `AuthorizationLayer` (Section 5).
///
/// Runs *after* [`crate::auth::auth_middleware`]. For block-scoped requests it
/// parses `X-Active-Block` (rejecting a missing/malformed header with `400`),
/// resolves the caller from extensions, and asks the policy. A panicking
/// policy MUST yield `403`, never `500`: the async policy call is isolated via
/// [`tokio::task::spawn`]; a panic surfaces as a [`tokio::task::JoinError`]
/// on `.await`, which `.unwrap_or(PolicyDecision::Deny)` maps to `Deny` → `403`.
pub async fn authorize_middleware(
    state: axum::extract::State<Arc<AuthorizationState>>,
    req: axum::http::Request<axum::body::Body>,
    next: Next,
) -> Response {
    let path = req.uri().path().to_string();
    // Documentation endpoints are public.
    if path.starts_with("/swagger-ui") || path.starts_with("/api-docs") {
        return next.run(req).await;
    }

    if matches!(requirement_for(&path), Requirement::Authenticated) {
        return next.run(req).await;
    }

    // Block-scoped: need a valid active block and an authenticated caller.
    // Parse `X-Active-Block` from the request parts (missing/malformed -> 400).
    let (mut parts, body) = req.into_parts();
    let active_block = match ActiveBlock::from_request_parts(&mut parts, &()).await {
        Ok(ab) => ab,
        Err(rej) => return rej.into_response(),
    };
    let current_user = parts.extensions.get::<CurrentUser>().cloned();
    let req = axum::http::Request::from_parts(parts, body);
    let current_user = match current_user {
        // Defensive: the auth layer should have populated this already.
        Some(u) => u,
        None => return AuthError::Unauthorized.into_response(),
    };

    let ctx = AuthContext {
        actor: current_user.sub,
        block_id: active_block.0,
        action: if req.method() == Method::GET {
            Action::Read
        } else {
            Action::Write
        },
    };

    // Panic-resistant evaluation (AC5): a panicking policy yields Deny → 403.
    let decision = tokio::task::spawn({
        let policy = state.policy.clone();
        let ctx = ctx.clone();
        async move { policy.authorize(&ctx).await }
    })
    .await
    .unwrap_or(PolicyDecision::Deny);

    let allowed = if state.enforce {
        matches!(decision, PolicyDecision::Allow)
    } else {
        if matches!(decision, PolicyDecision::Deny) {
            tracing::warn!(
                "AUTHZ(log-only): denied {} {} for {}",
                req.method(),
                path,
                ctx.actor
            );
        }
        true
    };

    if !allowed {
        // ADR-031: the authorization layer produces a problem document like
        // every other error source (single builder at the HTTP boundary).
        // `detail` is localized by the Fluent bundle (ADR-031 D5); no English
        // override here — the client branches on the `domain.forbidden` code.
        return crate::problems::problem(breakdown_core::error_registry::DOMAIN_FORBIDDEN)
            .build()
            .into_response();
    }

    next.run(req).await
}

#[cfg(test)]
mod authorization_test;