# backend/openapi.yaml
components:
  schemas:
...
    AuditEntry:
      description: |-
        One row of the audit journal: who (`actor`) did what (`event_type` on
        `entity_type`/`entity_id`) when (`occurred_at`), with the event `payload`.

        `series_id` is the tenant dimension prepared for per-`ProjectId` tenancy
        (decision 9.2) and is `NULL` in v1. `payload` is the raw event serialized
        as JSON (generic, so any context's events fit the same row).
      properties:
        actor:
          oneOf:
          - $ref: '#/components/schemas/UserId'
          - type: 'null'
        block_id:
          oneOf:
          - $ref: '#/components/schemas/BlockId'
          - type: 'null'
        entity_id:
          type: string
        entity_type:
          type: string
        event_type:
          type: string
        id:
          format: uuid
          type: string
        occurred_at:
          format: date-time
          type: string
        payload: {}
        series_id:
          format: uuid
          type:
          - string
          - 'null'
      required:
      - id
      - entity_type
      - entity_id
      - event_type
      - payload
      - occurred_at
      type: object
    BlockId:
      description: Opaque identifier for a `Block` aggregate.
      format: uuid
      type: string
...
    CreateBlockRequest:
      properties:
        end_date:
          format: date
          type:
          - string
          - 'null'
        number:
          format: int32
          type: integer
        season_id:
          $ref: '#/components/schemas/SeasonId'
        series_id:
          $ref: '#/components/schemas/ProjectId'
        start_date:
          format: date
          type:
          - string
          - 'null'
      required:
      - season_id
      - series_id
      - number
      type: object
...
    EpisodeId:
      description: Opaque identifier for an `Episode` aggregate.
      format: uuid
      type: string
...
    ProblemDetails:
      description: |-
        An RFC 9457 problem document (ADR-031 D1).

        Every error response (status ≥ 400) produced by the API is an instance of
        this document: `type` (derived from `code`), constant English `title`,
        `status`, stable `code`, localized `detail` (English until Tranche 3), and
        the `trace_id` extension for otel correlation.
      example:
        code: concurrency.version-mismatch
        detail: Version conflict
        extensions:
          current_version: 3
          expected_version: 2
        status: 409
        title: Version conflict
        trace_id: 4bf92f3577b34da6a3ce929d0e0e4736
        type: https://docs.breakdown.example/problems/scene.already-scheduled
      properties:
        code:
          description: Stable machine identity `{context}.{reason}` (the client contract).
          example: concurrency.version-mismatch
          type: string
        detail:
          description: Human-readable explanation; localized server-side (Tranche 3).
          example: Version conflict
          type: string
        extensions:
          additionalProperties: {}
          description: Declared S0/S1 extension fields, if any (ADR-031 D4).
          propertyNames:
            type: string
          type:
          - object
          - 'null'
        status:
          description: Canonical HTTP status of this problem.
          example: 409
          format: int32
          minimum: 0
          type: integer
        title:
          description: Constant English title (never localized; cacheable, spec-stable).
          type: string
        trace_id:
          description: W3C trace id of the request's otel span (support correlation).
          example: 4bf92f3577b34da6a3ce929d0e0e4736
          type: string
        type:
          description: Dereferencable documentation anchor, derived from the `code`.
          example: https://docs.breakdown.example/problems/scene.already-scheduled
          type: string
      required:
      - type
      - title
      - status
      - code
      - detail
      - trace_id
      type: object
    ProjectId:
      description: |-
        Opaque identifier for a `Project` — the tenant-level production container.

        `ProjectId` is an opaque UUIDv7 value type introduced by the
        `introduce-season-block-episode-hierarchy` change and renamed from
        `SeriesId` by issue #591 (ADR-035 D1/S1). The rename is **1:1**: the
        position, the UUIDv7 values and the tenant boundary are unchanged (ADR-035
        B1) — only the production-form-specific *word* is gone, so the container no
        longer claims to be a TV show.

        It remains a seam rather than an aggregate: every hierarchy entity
        (Season, Block, Episode) references it, but no `Project` aggregate exists
        yet. The children below it stay `Season`/`Block`/`Episode` until a second
        production form actually lands (ADR-035 D2/D3).

        **Wire compatibility (issue #591, layers 2 and 3 explicitly out of the
        rename).** The *type* is `#[serde(transparent)]` over `Uuid`, so a field
        typed `ProjectId` still serializes as a bare UUID string — the JSON value is
        unchanged by this rename. What deliberately keeps the old spelling is the
        **field name** on persisted payloads: event fields and read-model view
        fields carry `#[serde(rename = "series_id")]`, and the projection columns and
        OpenAPI fields keep `project_id` until the dedicated layer-3 change (a
        breaking ADR-021 `/v2` migration).
      format: uuid
      type: string
...
    SeasonId:
      description: Opaque identifier for a `Season` aggregate.
      format: uuid
      type: string
...
info:
  description: |-
    Costume scheduling API for Breakdown RS (ADR-006). Routes are mounted under the /v1 path prefix; the path version is bumped only on a breaking wire change (ADR-021).
    Error contract (ADR-031): every error response (status ≥ 400) is an RFC 9457 `application/problem+json` document carrying a stable `code` (`{context}.{reason}` kebab-case) and a `trace_id` extension. The full code registry is exposed under `x-code-registry`; per-code documentation lives at `docs/errors/`.
  license:
    identifier: AGPL-3.0
    name: AGPL-3.0
  title: Breakdown RS API
  version: v1
openapi: 3.1.0
paths:
...
  /v1/audit:
    get:
      description: Audit journal entries of a series, newest first. Requires the project_id query parameter (400 otherwise).
      operationId: get_audit_history
      parameters:
      - in: query
        name: limit
        required: false
        schema:
          default: 50
          format: int64
          type:
          - integer
          - 'null'
      - in: query
        name: offset
        required: false
        schema:
          default: 0
          format: int64
          type:
          - integer
          - 'null'
      - in: query
        name: episode_id
        required: false
        schema:
          oneOf:
          - $ref: '#/components/schemas/EpisodeId'
          - type: 'null'
      - in: query
        name: season_id
        required: false
        schema:
          oneOf:
          - $ref: '#/components/schemas/SeasonId'
          - type: 'null'
      - in: query
        name: series_id
        required: false
        schema:
          oneOf:
          - $ref: '#/components/schemas/ProjectId'
          - type: 'null'
      responses:
        '200':
          content:
            application/json:
              schema:
                items:
                  $ref: '#/components/schemas/AuditEntry'
                type: array
          description: Audit journal entries of the series, newest first
        '400':
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
          description: Missing project_id query parameter
        '403':
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
          description: Caller holds no active membership in any block of the queried series
        '422':
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
          description: Validation error
      tags:
      - handlers
  /v1/blocks:
    get:
      description: Lists the blocks of a season. Requires the season_id query parameter (400 otherwise).
      operationId: list_blocks
      parameters:
      - in: query
        name: limit
        required: false
        schema:
          default: 50
          format: int64
          type:
          - integer
          - 'null'
      - in: query
        name: offset
        required: false
        schema:
          default: 0
          format: int64
          type:
          - integer
          - 'null'
      - in: query
        name: episode_id
        required: false
        schema:
          oneOf:
          - $ref: '#/components/schemas/EpisodeId'
          - type: 'null'
      - in: query
        name: season_id
        required: false
        schema:
          oneOf:
          - $ref: '#/components/schemas/SeasonId'
          - type: 'null'
      - in: query
        name: series_id
        required: false
        schema:
          oneOf:
          - $ref: '#/components/schemas/ProjectId'
          - type: 'null'
      responses:
        '200':
          content:
            application/json:
              schema:
                items:
                  $ref: '#/components/schemas/BlockView'
                type: array
        '400':
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
          description: Missing season_id query parameter
      tags:
      - handlers
    post:
      operationId: create_block
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateBlockRequest'
        required: true
      responses:
        '201':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/IdVersionResponse'
          description: Block created
        '409':
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
          description: Conflict
        '422':
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
          description: Validation error
      tags:
      - handlers
...
  /v1/episodes:
    get:
      description: Lists episodes of a series, or of a single block when block_id is given. Requires project_id unless block_id is present (400 otherwise).
      operationId: list_episodes
      parameters:
      - in: query
        name: limit
        required: false
        schema:
          default: 50
          format: int64
          type:
          - integer
          - 'null'
      - in: query
        name: offset
        required: false
        schema:
          default: 0
          format: int64
          type:
          - integer
          - 'null'
      - in: query
        name: series_id
        required: false
        schema:
          oneOf:
          - $ref: '#/components/schemas/ProjectId'
          - type: 'null'
      - in: query
        name: block_id
        required: false
        schema:
          oneOf:
          - $ref: '#/components/schemas/BlockId'
          - type: 'null'
      responses:
        '200':
          content:
            application/json:
              schema:
                items:
                  $ref: '#/components/schemas/EpisodeView'
                type: array
        '400':
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
          description: Missing project_id query parameter
      tags:
      - handlers
    post:
      operationId: create_episode
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateEpisodeRequest'
        required: true
      responses:
        '201':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/IdVersionResponse'
          description: Episode created
        '409':
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
          description: Conflict
        '422':
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
          description: Validation error
      tags:
      - handlers
...
  /v1/seasons:
    get:
      description: Lists seasons, optionally narrowed to one series via project_id. The table stays small, so — unlike the episode/scene lists — no scope parameter is required.
      operationId: list_seasons
      parameters:
      - in: query
        name: limit
        required: false
        schema:
          default: 50
          format: int64
          minimum: 0
          type:
          - integer
          - 'null'
      - in: query
        name: offset
        required: false
        schema:
          default: 0
          format: int64
          minimum: 0
          type:
          - integer
          - 'null'
      - in: query
        name: series_id
        required: false
        schema:
          oneOf:
          - $ref: '#/components/schemas/ProjectId'
          - type: 'null'
      - description: |-
          Issue #533: archived seasons are hidden by default; the explicit
          opt-in returns them (read-only, still locked for writes).
        in: query
        name: include_archived
        required: false
        schema:
          default: false
          type:
          - boolean
          - 'null'
      responses:
        '200':
          content:
            application/json:
              schema:
                items:
                  $ref: '#/components/schemas/SeasonView'
                type: array
        '400':
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
          description: Negative limit or offset
        '409':
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
          description: Projection store failure
      tags:
      - handlers
    post:
      operationId: create_season
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateSeasonRequest'
        required: true
      responses:
        '201':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/IdVersionResponse'
          description: Season created
        '409':
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
          description: Conflict
        '422':
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
          description: Validation error
      tags:
      - handlers
...
  /v1/seasons/{id}/membership:
    get:
      operationId: get_season_membership
      parameters:
      - description: Season id
        in: path
        name: id
        required: true
        schema:
          format: uuid
          type: string
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SeasonMembershipDto'
          description: Membership of the authenticated caller in the season
        '401':
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
          description: Authentication required
        '404':
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
          description: Season not found
      tags:
      - handlers
...
  /v1/series/{id}/membership:
    get:
      operationId: get_series_membership
      parameters:
      - description: Series id
        in: path
        name: id
        required: true
        schema:
          format: uuid
          type: string
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SeriesMembershipDto'
          description: Membership of the authenticated caller in the series
        '401':
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
          description: Authentication required
      tags:
      - handlers
...